Refactor login page and update dependencies
