+++ app/login/page.tsx
@@ -5,143 +5,22 @@
  * No database, no complex auth - just checks against ADMIN_PASSWORD env variable.
  * 
  * SEO:
- * - Includes noindex/nofollow meta tags to prevent search engine indexing
+ * - Includes noindex/nofollow via metadata
  * - Also blocked in robots.txt
  */
 
-'use client';
+import { Metadata } from 'next';
+import LoginForm from './LoginForm';
 
-import { useState } from 'react';
-import { useSearchParams } from 'next/navigation';
-import { Button } from '@/app/components/ui/button';
-import { Input } from '@/app/components/ui/input';
-import { Label } from '@/app/components/ui/label';
-import { Shield, Lock } from 'lucide-react';
+// Metadata for SEO - prevents indexing
+export const metadata: Metadata = {
+  title: 'Admin Login',
+  robots: {
+    index: false,
+    follow: false,
+  },
+};
 
 export default function LoginPage() {
-  const [password, setPassword] = useState('');
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-
-  const searchParams = useSearchParams();
-  const callbackUrl = searchParams.get('callbackUrl') || '/pages/admin/integrations';
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setLoading(true);
-    setError(null);
-
-    try {
-      const res = await fetch('/api/auth/login', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ password }),
-      });
-
-      const data = await res.json();
-
-      if (res.ok && data.success) {
-        // Redirect to callback URL on success
-        window.location.href = callbackUrl;
-      } else {
-        setError(data.error || 'Invalid password');
-        setLoading(false);
-      }
-    } catch (err) {
-      setError('An error occurred. Please try again.');
-      setLoading(false);
-    }
-  };
-
-  return (
-    <>
-      {/* SEO: Prevent indexing of login page */}
-      <head>
-        <meta name="robots" content="noindex, nofollow" />
-        <title>Admin Login</title>
-      </head>
-
-      <main className="min-h-screen flex items-center justify-center p-6 bg-gradient-to-br from-slate-50 to-slate-100 dark:from-slate-950 dark:to-slate-900">
-        <div className="w-full max-w-md">
-          <div className="bg-white dark:bg-slate-800 shadow-2xl rounded-2xl p-8 border border-slate-200 dark:border-slate-700">
-            {/* Header */}
-            <div className="text-center mb-8">
-              <div className="inline-flex p-3 bg-primary/10 rounded-full mb-4">
-                <Shield className="h-8 w-8 text-primary" />
-              </div>
-              <h1 className="text-3xl font-bold text-slate-900 dark:text-white mb-2">
-                Admin Access
-              </h1>
-              <p className="text-slate-600 dark:text-slate-400">
-                Enter your password to continue
-              </p>
-            </div>
-
-            {/* Error Message */}
-            {error && (
-              <div className="mb-6 p-4 bg-red-50 dark:bg-red-900/20 border border-red-200 dark:border-red-800 rounded-lg">
-                <p className="text-sm text-red-800 dark:text-red-300">{error}</p>
-              </div>
-            )}
-
-            {/* Login Form */}
-            <form onSubmit={handleSubmit} className="space-y-5">
-              <div className="space-y-2">
-                <Label htmlFor="password" className="text-slate-700 dark:text-slate-300 flex items-center gap-2">
-                  <Lock className="h-4 w-4" />
-                  Password
-                </Label>
-                <Input
-                  id="password"
-                  type="password"
-                  placeholder="Enter admin password"
-                  value={password}
-                  onChange={(e) => setPassword(e.target.value)}
-                  disabled={loading}
-                  required
-                  className="w-full"
-                  autoComplete="current-password"
-                  autoFocus
-                />
-              </div>
-
-              <Button
-                type="submit"
-                disabled={loading}
-                className="w-full h-11 text-base font-medium"
-              >
-                {loading ? (
-                  <span className="flex items-center gap-2">
-                    <svg
-                      className="animate-spin h-5 w-5"
-                      xmlns="http://www.w3.org/2000/svg"
-                      fill="none"
-                      viewBox="0 0 24 24"
-                    >
-                      <circle
-                        className="opacity-25"
-                        cx="12"
-                        cy="12"
-                        r="10"
-                        stroke="currentColor"
-                        strokeWidth="4"
-                      />
-                      <path
-                        className="opacity-75"
-                        fill="currentColor"
-                        d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"
-                      />
-                    </svg>
-                    Verifying...
-                  </span>
-                ) : (
-                  'Sign In'
-                )}
-              </Button>
-            </form>
-          </div>
-        </div>
-      </main>
-    </>
-  );
+  return <LoginForm />;
 }
