+++ package.json
@@ -31,7 +31,7 @@
     "@radix-ui/react-switch": "^1.2.6",
     "@radix-ui/react-tabs": "^1.1.13",
     "@react-three/drei": "^10.0.4",
-    "@react-three/fiber": "9.0.0-alpha.8",
+    "@react-three/fiber": "^9.0.0",
     "@react-three/postprocessing": "^3.0.4",
     "@tabler/icons-react": "^3.35.0",
     "@tsparticles/engine": "^3.9.1",
