+++ public/sitemap.xml
@@ -3,49 +3,49 @@
   
     <url>
       <loc>https://enterprisehero.com/</loc>
-      <lastmod>2025-10-06T16:09:48.451Z</lastmod>
+      <lastmod>2025-10-06T16:14:02.901Z</lastmod>
       <changefreq>weekly</changefreq>
       <priority>1</priority>
     </url>
   
     <url>
       <loc>https://enterprisehero.com/pages/about</loc>
-      <lastmod>2025-10-06T16:09:48.452Z</lastmod>
+      <lastmod>2025-10-06T16:14:02.901Z</lastmod>
       <changefreq>monthly</changefreq>
       <priority>0.8</priority>
     </url>
   
     <url>
       <loc>https://enterprisehero.com/pages/services</loc>
-      <lastmod>2025-10-06T16:09:48.452Z</lastmod>
+      <lastmod>2025-10-06T16:14:02.901Z</lastmod>
       <changefreq>monthly</changefreq>
       <priority>0.8</priority>
     </url>
   
     <url>
       <loc>https://enterprisehero.com/pages/portfolio</loc>
-      <lastmod>2025-10-06T16:09:48.452Z</lastmod>
+      <lastmod>2025-10-06T16:14:02.901Z</lastmod>
       <changefreq>weekly</changefreq>
       <priority>0.9</priority>
     </url>
   
     <url>
       <loc>https://enterprisehero.com/pages/blog</loc>
-      <lastmod>2025-10-06T16:09:48.452Z</lastmod>
+      <lastmod>2025-10-06T16:14:02.901Z</lastmod>
       <changefreq>daily</changefreq>
       <priority>0.9</priority>
     </url>
   
     <url>
       <loc>https://enterprisehero.com/pages/resources</loc>
-      <lastmod>2025-10-06T16:09:48.452Z</lastmod>
+      <lastmod>2025-10-06T16:14:02.901Z</lastmod>
       <changefreq>weekly</changefreq>
       <priority>0.8</priority>
     </url>
   
     <url>
       <loc>https://enterprisehero.com/pages/contact</loc>
-      <lastmod>2025-10-06T16:09:48.452Z</lastmod>
+      <lastmod>2025-10-06T16:14:02.901Z</lastmod>
       <changefreq>monthly</changefreq>
       <priority>0.7</priority>
     </url>
